src/docker/BI/requirements.txt
pandas
Pillow
plotly
dash
jupyter-dash
numpy
statsmodels
